Merge SRD changes from amrex PR#3491
Fix a bug in StateRedistribution which occurred when the EB was
near a fine-fine boundary. In this case we weren't calculating
certain arrays in a wide enough band of ghost cells. This has been
fixed by enlarging two of the arrays and calculating them on the
correct number of ghost cells.